Move a couple more files
